+++ Runtime/Animator/AnimatorParameterResetter.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using UnityEngine;
 
-namespace Mane.StateMachine
+namespace Mane.AnimatorStateMachine
 {
     public class AnimatorParameterResetter : StateMachineBehaviour
     {